+++ NN-Toolkit/src/Layers/MaxPooling2D.cpp
@@ -10,8 +10,8 @@ namespace NN {
             int stride,
             float (*activation) (float, int, float),
             bool bias,
-            int rand_from,
-            int rand_to,
+            float rand_from,
+            float rand_to,
             float hyperparameter
         ) : f(pool_size),
             p(padding),
